python_rag/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
pydantic>=2.8,<3
python-dotenv>=1,<2
python-multipart>=0.0.9,<1
PyMySQL>=1.1,<2
redis>=5,<6
celery>=5.4,<6
requests>=2.32,<3
numpy>=1.26,<3
faiss-cpu>=1.8,<2
sentence-transformers>=3,<4
torch>=2.3,<3
